Add files via upload
This analysis uses hypothesis testing for the population mean as well as hypothesis testing for the difference between two population means using a dataset from the NBA of the Bulls from the seasons 1996-1998 and the Hawks from 2013-2015. The code for this project is written in Python. Visual files are saved as HTML.